requirements.txt
FastAPI==0.104.0
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
